backend/requirements.txt
alembic==1.8.1
astroid==2.12.11
bidict==0.22.0
click==8.1.3
dill==0.3.5.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.1
greenlet==1.1.3.post0
h11==0.14.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
pkg_resources==0.0.0
platformdirs==2.5.2
psycopg2==2.9.4
pylint==2.15.4
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
simple-websocket==0.8.1
six==1.16.0
SQLAlchemy==1.4.42
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.9.0
